Deletes all entries before inserting new collections
* There can be instances when records get updated in SOLR and collection info
for them is deleted. Therefore, we need to make sure our collections in the
entries table are up to date. Since, it is more tedious to go through every
row in the table and check if entry still exists in the db table, we are
deleting all entries in the entries table with a single SQL query.