CI Updates (#1350)
* Use CircleCIs docker images

- The old images have been deprecated for over a year, so start using the
new ones. Additionally, a CircleCI orb is necessary to install browsers
for feature tests.

- Updates Redis and Mysql. The old versions, '4', and '5.7.22' are not
available in Circle's new images. This upgrades to the latest for each,
'7.0', and '8.0'.

- Adds SQLite headers. Installing gems fails because the SQLite dev
package is missing.

* Remove Yarn commands from CI

There is no yarn.lock, so CI could not retrieve a cached version. We do
have a package.json, so I think the lock file is missing.

* Add yarn.lock

---------

Co-authored-by: Ayoub Belemlih <13107510+abelemlih@users.noreply.github.com>